Ignore camera pitch. Don't normalize controller.
Before, it would slow you down if you looked at the ground While
FromYaw(rot.Yaw()) is ugly, rot is quaternions, and I need a single yaw
value. I'd like for there to be a rot.WithPitch or something similar,
even if it just does Rotation.FromAngles(rot.Angles().WithPitch()).

Also, decided not to normalize the controller, so if you press lightly,
your character walks.
It kinda sucks that with the amount of friction the floor has, there's a
very fine line between walking and running. The good news is that
there's a massive deadzone due to that, so you can safely play the game
on your keyboard with the controller plugged in, and it won't throw you
off the edge. I think in the previous commit, it would do that. :(

That reminds me, I gotta make it more strict with the NearlyZero, just
in case someone is holding Forwards and Backwards on their keyboard, and
their controller is plugged in.